Add logging for editpos and patch.
